+++ grafana_snapshots/constants.py
@@ -1,5 +1,5 @@
 
 PKG_NAME = 'grafana-snapshots'
-PKG_VERSION = '0.3.1'
+PKG_VERSION = '0.3.2'
 CONFIG_NAME = 'conf/grafana-snapshots.yml'